database/requirements.txt
pymongo 
python-dotenv
ipykernel
pydantic